[fix] router 삭제
